[firebase_ml_vision] Prepare example app to include Live Camera Preview demo and Material Barcode Scanner demo (#1713)
